refactor: break up hashing.go to reduce dependencies. According to Go practice, too much DRY leads to too many dependencies, thus some duplication is allowed. leftPadByteSlice, buildOracleQueryID, buildContractID moved to transactions.go Namehash moved to helpers.go randomBytes duplicated in generateCommitmentID and moved to keystore.go:KeystoreSeal uuidV4 moved to keystore.go generateCommitmentID/computeCommitmentID moved to helpers.go buildRLPMessage, buildIDTag, readIDTag moved to transactions.go
